refactor: restructure from main to separate files
